common: Move wwinit functions library file to libexec directory
The sysconf directory (traditionally /etc) should only contain
host-specific configuration files. As the functions library file is a
source-able script with no host-specific configuration parameters, it
can be considered a read-only general-use file and would be better
placed in the libexec directory.

The wwinit program has an established executable directory where
modules may add scripts as necessary for initialization. The wwinit
utility specifically searches that directory for files that end with
".init" in the filename, and thus a file with the name of "functions"
would be ignored by wwinit unless explicitly referenced. Thus, putting
the functions library file in the wwinit libexec directory should not
cause any conflict.

Signed-off-by: John L. Jolly <jjolly@suse.com>