provision: Only check for libtirpc if busybox will be built
The build system would check for the presense of the libtirpc package
even if a local busybox would be used and the busybox build process
would be skipped.

This change causes the test for libtirpc to only be made if the internal
busybox build will be performed. Otherwise, the test is skipped. If a
local busybox will be used, then it is expected to already have all
dependent libraries installed.

Signed-off-by: John L. Jolly <jjolly@suse.com>