provision: Add check for libraries required by parted
The parted build requires libuuid and device mapper to be installed.
When the development packages for these requirements are not available,
the parted build fails.

This fix adds a package check for uuid and device-mapper into the
configure.ac for the provision module. This check is only made if the
internal build of parted will be performed. If a local copy of parted
will be used, then these checks will be skipped.

Signed-off-by: John L. Jolly <jjolly@suse.com>