provision: Add check for liblzma for undionly ipxe
The undionly ipxe build requires liblzma development files to be
installed. If these files cannot be found, the undionly ipxe build
fails.

This fix adds a check for liblzma development files, but only when the
undionly ipxe image will be built internally. If a local copy of the
undionly ipxe image will be used, then this check is skipped.

Signed-off-by: John L. Jolly <jjolly@suse.com>